requirements.txt
pandas==1.3.4
Pillow==8.4.0
streamlit==1.1.0
